requirements.txt
Django==1.7.2
Pillow==2.7.0
PyJWT==0.4.1
braintree==3.10.0
django-crispy-forms==1.4.0
django-markdown-deux==1.0.5
django-pagedown==0.1.1
markdown2==2.3.1
oauthlib==0.7.2
python-openid==2.2.5
python-social-auth==0.2.1
requests==2.5.1
requests-oauthlib==0.4.2
six==1.9.0
wsgiref==0.1.2
